finding_missing_person_using_euclidian_loss/requirements.txt
streamlit
pymongo
dnspython
pillow